Refactor: 해당 년도, 월의 날짜를 가져오는 함수 재사용
