+++ src/pages/games/DateQuiz.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 import { Container, Desc, Body } from '../../components/games/DateQuiz';
 import { GameProps } from '../../routes/gameRouter.tsx';
+import { getMonthDate } from '../../modules/getMonthDate.ts';
 
 export default function DateQuiz({
   gameData,
@@ -14,43 +15,10 @@ export default function DateQuiz({
   const YEAR = DATE.getFullYear();
   const MONTH = DATE.getMonth() + 1;
   const TODAY = DATE.getDate();
-
   const [totalDate, setTotalDate] = useState<number[]>([]);
 
-  const changeDate = (year: number, month: number) => {
-    // 이전 날짜
-    const PVLastDate = new Date(year, month - 1, 0).getDate();
-    const PVLastDay = new Date(year, month - 1, 0).getDay();
-
-    // 다음 날짜
-    const ThisLasyDay = new Date(year, month, 0).getDay();
-    const ThisLasyDate = new Date(year, month, 0).getDate();
-
-    // 이전 날짜 만들기
-    let PVLD = [];
-    if (PVLastDay !== 6) {
-      for (let i = 0; i < PVLastDay + 1; i++) {
-        PVLD.unshift(PVLastDate - i);
-      }
-    }
-
-    // 다음 날짜 만들기
-    let TLD = [];
-    for (let i = 1; i < 7 - ThisLasyDay; i++) {
-      if (i === 0) {
-        return TLD;
-      }
-      TLD.push(i);
-    }
-
-    // 현재날짜
-    let TD = [];
-    TD = [...Array(ThisLasyDate + 1).keys()].slice(1);
-    return PVLD.concat(TD, TLD);
-  };
-
   useEffect(() => {
-    setTotalDate(changeDate(YEAR, MONTH));
+    setTotalDate(getMonthDate(YEAR, MONTH));
   }, []);
 
   return (
