+++ src/pages/games/DayOfWeek.tsx
@@ -15,6 +15,7 @@ import {
   DayContainer,
 } from '../../components/games/DayOfWeek.tsx';
 import { getRandomFloat } from '../../utils/random.ts';
+import { getMonthDate } from '../../modules/getMonthDate.ts';
 
 function DayOfWeek({
   gameData,
@@ -53,38 +54,6 @@ function DayOfWeek({
   };
 
   const currentWeek = getWeek(DATE);
-
-  const changeDate = (year: number, month: number) => {
-    // 이전 날짜
-    const PVLastDate = new Date(year, month - 1, 0).getDate();
-    const PVLastDay = new Date(year, month - 1, 0).getDay();
-
-    // 다음 날짜
-    const ThisLasyDay = new Date(year, month, 0).getDay();
-    const ThisLasyDate = new Date(year, month, 0).getDate();
-
-    // 이전 날짜 만들기
-    let PVLD = [];
-    if (PVLastDay !== 6) {
-      for (let i = 0; i < PVLastDay + 1; i++) {
-        PVLD.unshift(PVLastDate - i);
-      }
-    }
-
-    // 다음 날짜 만들기
-    let TLD = [];
-    for (let i = 1; i < 7 - ThisLasyDay; i++) {
-      if (i === 0) {
-        return TLD;
-      }
-      TLD.push(i);
-    }
-
-    // 현재날짜
-    let TD = [];
-    TD = [...Array(ThisLasyDate + 1).keys()].slice(1);
-    return PVLD.concat(TD, TLD);
-  };
 
   const splitArrayIntoChunks = (arr: number[], chunkSize: number) => {
     const chunks = [];
@@ -169,7 +138,7 @@ function DayOfWeek({
   };
 
   useEffect(() => {
-    const totalDate = changeDate(YEAR, MONTH);
+    const totalDate = getMonthDate(YEAR, MONTH);
     setTotalDate(totalDate);
     generateRandomDate(totalDate);
   }, []);
